backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
pydantic==2.6.3
pydantic-settings==2.2.1
httpx==0.24.1
python-dotenv==1.0.1
supabase==1.2.0
pytest==8.0.2
pytest-cov==4.1.0
